+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 redis extension Change Log
 2.0.18 under development
 ------------------------
 
-- no changes in this release.
+- Bug #247: `Cache::getValue()` now returns `false` in case of missing key (rhertogh)
 
 
 2.0.17 January 11, 2022