+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Yii Framework 2 redis extension Change Log
 ==========================================
 
+2.0.19 under development
+------------------------
+
+- no changes in this release.
+
+
 2.0.18 September 04, 2022
 -------------------------